refactor: apply recommendations and fix critical bugs
- fix(physics): prevent double knockback side-effects on generic entities
- fix(combat): unify damage amounts to float across the combat pipeline
- fix(player): move mutable input state flags to __init__
- fix(core): remove duplicate pytest dependencies in pyproject.toml
- refactor(states): extract magic numbers to centralized settings.py
- refactor(types): modernize typing annotations in state_machine & sprites
- refactor(core): replace print statements with logging in game.py
- chore(vcs): stop ignoring the assets/ directory in .gitignore
- test: add basic unit tests for Player and StateMachine
- style: rename _is_wall_sliding and _dash_requested to public properties

+++ src/combat/combat_component.py
@@ -287,7 +287,7 @@ class CombatComponent:
 
     def take_damage(
         self,
-        amount: int,
+        amount: float,
         source_center_x: float | None = None,
         knockback: KnockbackConfig | None = None,
     ) -> None:
@@ -295,7 +295,7 @@ class CombatComponent:
 
         Parameters
         ----------
-        amount : int
+        amount : float
             Hit points to subtract.
         source_center_x : float | None
             X centre of the damage source for knockback direction.
@@ -440,7 +440,7 @@ class NullCombatComponent:
 
     def take_damage(
         self,
-        amount: int,
+        amount: float,
         source_center_x: float | None = None,
         knockback: KnockbackConfig | None = None,
     ) -> None:
